pyproject.toml
[tool.poetry]
name = "master-bot"
version = "0.1.0"
description = ""
authors = ["Timofei Balduev <balduev.desco@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiosqlite = ">=0.19.0,<0.20.0"
python-dotenv = ">=1.0.0,<1.1.0"
python-telegram-bot = ">=20.4,<21.0"
alembic = "^1.12.0"
asyncpg = "^0.28.0"
psycopg = "^3.1.12"
psycopg-binary = "^3.1.12"
sqlalchemy = "^2.0.22"
pydantic-settings = "^2.1.0"
aiogram = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
